Make mocking helpers independent of vscode
